[Refactor]:[Split bookCard component] Changes to be committed: modified:   src/components/bookCard.js new file:   src/components/cardPriority.js
